% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_sum_TreatyLoss100Participation.R
\name{apac}
\alias{apac}
\title{APAC Industry Loss Numbers}
\arguments{
\item{event_id}{The EventID to search for}

\item{model_code}{The ModelCode to search for}

\item{insured_csv}{Path to the insured CSV file, remember to include ""}

\item{insurable_csv}{Path to the insurable CSV file, remember to include ""}

\item{output_csv}{Path to the output CSV file, remember to include ""}
}
\value{
A data frame with the Event ID, Model Code, Country Code, Sum of Insured and Insurable Loss
}
\description{
Search and summarise the Industry Loss Numbers based on a given EventID and ModelCode
}
\examples{
apac(267483,18,"insured.csv file path","insurable.csv file path", "output.csv file path")
}
